rearrange outcome package
outcome.go renamed to exit.go
allocations and tests split to separate files
single asset exits and tests split to separate files
test data split to a separate file
