+++ channel/state/outcome/single-asset-exit_test.go
@@ -1,0 +1,1 @@
+package outcome
